check in secrets master key
This is the simplest way to get the logjam docker
container running again. It’s not a security issue
as we have no sensitive information at all in the
open source version of the app. 

For a production installation of logjam that adds
real secrets, the master key should be removed and
new credentials should be generated.